Fix path traversal issue.
The code was implementing a path traversal check based on the detection
of '..' in directory names and file names. This is sufficient to protect
against path traversal attacks using relative paths but insuficcient for
attacks using absolute paths.

This is due to the fact that the second argument of os.path.join() takes
precedence if it starts with '/':

>>> print(os.path.join('outdir', '/tmp/hacked'))
/tmp/hacked